Add one-hot encoded action space representation
with fixed size.

Also rename utils to hive_utils.

Also create train_test.py which will be responsible for testing the
train of the alphaGo AI.

Format of the fixed size action space is the following:

- The first <len>-1 numbers (where <len> is the number of pieces)
represents the placement of the first piece. These moves are only
allowed in the first turn of the player. In that case, you can place
every piece except the queen. That's why we need only <len>-1 fields for
that.

In the second player's first turn, even though he can place his bugs to
6 different locations, it can still be signed as <len>-1 possible moves
since the resulting boards are identical.

- The next <len>*(<len>-1)*6 numbers represents the pieces that can be
placed down next to an allied bug. They are all shut to 0 in case the
player has only his first turn. Each bug can be placed next to <len>-1
allied bugs. Each of those bugs have 6 sides where the bug can be placed
to.

- The remaining numbers represents the movements. Each bug can have a
fixed amount of moves. E.g. I limited the ant's movement space to 50. It
can be changed if it turns out to be not enough.